requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alpha-vantage==2.3.1
amqp==5.0.9
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.2
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-heroku==0.0.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
frozenlist==1.3.0
gunicorn==19.7.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
multidict==6.0.2
openapi-codec==1.3.2
pandas
django-environ
prompt-toolkit==3.0.26
psycopg2-binary==2.9.3
pytz==2021.3
PyYAML==6.0
redis==2.10.3
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
whitenoise==3.3.1
yarl==1.7.2
